NXP-32165: Make GridFS Blob Store delete idempotent

+++ modules/core/nuxeo-core-storage-mongodb/src/main/java/org/nuxeo/ecm/core/storage/mongodb/blob/GridFSBlobStore.java
@@ -159,7 +159,9 @@ public class GridFSBlobStore extends AbstractBlobStore {
     @Override
     public void deleteBlob(String key) {
         GridFSFile dbFile = getGridFSFile(key);
-        getGridFSBucket().delete(dbFile.getId());
+        if (dbFile != null) {
+            getGridFSBucket().delete(dbFile.getId());
+        }
     }
 
     @Override
